Separate some helpers into utils/index.ts
